+++ wiki-fe/src/app/(pages)/(login)/cadastrar/page.jsx
@@ -1,86 +1,13 @@
 "use client"
-
-import React, { useState } from 'react';
-import "@/styles/signup.css";
+import Footer from '@/components/Footer';
+import SignUp from '@/components/SignUp';
 
 export default function SignUpPage() {
-  const [nome, setNome] = useState('');
-  const [email, setEmail] = useState('');
-  const [username, setUsername] = useState('');
-  const [password, setPassword] = useState('');
-  const [termsAccepted, setTermsAccepted] = useState(false);
-
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-    const formData = { username, password, nome, email, termsAccepted };
-
-    try {
-      const response = await fetch('http://127.0.0.1:8000/create_user', { 
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(formData), 
-      });
-    } catch (error) {
-      console.log(error);
-    }
-};
-  
   return (
-    <div className="signup-container">
-      <h2>Cadastrar</h2>
-      <form onSubmit={handleSubmit}>
-        <div>
-          <label htmlFor="username">Usuário:</label>
-          <input
-            type="text"
-            id="username"
-            value={username}
-            onChange={(e) => setUsername(e.target.value)}
-            required
-          />
-        </div>
-        <div>
-          <label htmlFor="password">Senha:</label>
-          <input
-            type="password"
-            id="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            required
-          />
-        </div>
-        <div>
-          <label htmlFor="nome">Nome:</label>
-          <input
-            type="text"
-            id="nome"
-            value={nome}
-            onChange={(e) => setNome(e.target.value)}
-            required
-          />
-        </div>
-        <div>
-          <label htmlFor="email">Email:</label>
-          <input
-            type="email"
-            id="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-          />
-        </div>
-        <div className="terms-container">
-          <input
-            type="checkbox"
-            id="terms"
-            checked={termsAccepted}
-            onChange={(e) => setTermsAccepted(e.target.checked)}
-            required
-          />
-          <label htmlFor="terms">Eu aceito os termos de uso</label>
-        </div>
-        <button type="submit">Cadastrar</button>
-      </form>
-    </div>
+    <>
+      <SignUp />
+
+      <Footer />
+    </>
   );
 }
