+++ wiki-fe/src/app/layout.jsx
@@ -1,6 +1,5 @@
 import '@/styles/layout.css'
 
-
 export const metadata = {
   title: 'Wiki Dengue',
 }
@@ -9,7 +8,7 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body>
-        {children}
+          {children}
       </body>
     </html>
   )
