+++ wiki-fe/src/app/page.jsx
@@ -1,8 +1,8 @@
-import Content from "@/components/Content";
-import loremIpsum from "@/assets/json/loremIpsum.json";
 import Body from "@/components/Body";
 import Header from "@/components/Header";
 import Footer from "@/components/Footer";
+import Content from "@/components/Content";
+import loremIpsum from "@/assets/json/loremIpsum.json";
 
 export default function Page() {
   return (
